metabase: Reuse table creation everywhere
